requirements.txt
# Airflow core
apache-airflow>=2.8.0

# Database dependencies
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0

# Web scraping and downloads
requests>=2.28.0
beautifulsoup4>=4.11.0
urllib3>=1.26.0
tqdm>=4.64.0

# Retry logic
tenacity>=8.2.0

# Environment variables
python-dotenv>=0.19.0

# YAML configuration
pyyaml>=6.0

# Additional Airflow providers (optional)
apache-airflow-providers-postgres>=5.0.0
apache-airflow-providers-http>=4.0.0 